admin updated, client admin created + purchase mod

+++ admin/includes/view_all_books.php
@@ -1,3 +1,5 @@
+<!-- <form action="" method="post"> -->
+  <div class="books-page">
 <table border="1" >
   <thead>
     <tr>
@@ -10,6 +12,11 @@
       <th>Tags</th>
       <th>Image</th>
       <th>Status</th>
+      <th>Comments</th>
+      <th>Date</th>
+      <th>View Book</th>
+      <th>Edit</th>
+      <th>Delete</th>
     </tr>
   </thead>
   <tbody>
@@ -28,7 +35,9 @@
         $book_author = $row['book_author'];
         $book_image = $row['book_image'];
         $book_tags = $row['book_tags'];
-        $book_status = $row['book_status'];
+        $book_status = $row['book_status']; 
+        $upload_date = $row['upload_date']; 
+        $book_comment_count = $row['book_comment_count']; 
 
         echo "<tr>";
         echo "<td>{$book_id}</td>";
@@ -52,8 +61,11 @@
         echo "<td>{$book_tags}</td>";
         echo "<td><img width='50' height='50' src='../images/$book_image' alt='{$book_name}'></td>";
         echo "<td>{$book_status}</td>";
+        echo "<td>{$book_comment_count}</td>";
+        echo "<td>{$upload_date}</td>";
+        echo "<td><a href='../single-book.php?b_id={$book_id}'>View Book</a></td>";
         echo "<td><a href='books.php?source=edit_book&b_id={$book_id}'>Edit</a></td>";
-        echo "<td><a href='books.php?delete={$book_id}'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='books.php?delete={$book_id}'>Delete</a></td>";
         echo "</tr>";
 
       }
@@ -63,5 +75,6 @@
     
   </tbody>
 </table>
-
+<!-- </form> -->
+</div>
 <?php delete_book(); ?>
